+++ triedb/firewood/database.go
@@ -26,6 +26,13 @@ import (
 	"github.com/ava-labs/libevm/triedb/database"
 )
 
+const (
+	// Directory where all Firewood state lives.
+	firewoodDir          = "firewood"
+	firewoodFileName     = "firewood.db"
+	firewoodRootStoreDir = "root_store"
+)
+
 var (
 	_ proposable = (*ffi.Database)(nil)
 	_ proposable = (*ffi.Proposal)(nil)
@@ -62,11 +69,12 @@ type ProposalContext struct {
 }
 
 type Config struct {
-	FilePath             string
+	ChainDataDir         string
 	CleanCacheSize       int  // Size of the clean cache in bytes
 	FreeListCacheEntries uint // Number of free list entries to cache
-	Revisions            uint
+	Revisions            uint // Number of revisions to keep in memory (must be >= 2)
 	ReadCacheStrategy    ffi.CacheStrategy
+	ArchiveMode          bool
 }
 
 // Note that `FilePath` is not specified, and must always be set by the user.
@@ -100,15 +108,23 @@ type Database struct {
 // New creates a new Firewood database with the given disk database and configuration.
 // Any error during creation will cause the program to exit.
 func New(config Config) (*Database, error) {
-	if err := validatePath(config.FilePath); err != nil {
+	firewoodDir := filepath.Join(config.ChainDataDir, firewoodDir)
+	filePath := filepath.Join(firewoodDir, firewoodFileName)
+	if err := validatePath(filePath); err != nil {
 		return nil, err
 	}
 
-	fw, err := ffi.New(config.FilePath, &ffi.Config{
+	var rootStoreDir string
+	if config.ArchiveMode {
+		rootStoreDir = filepath.Join(firewoodDir, firewoodRootStoreDir)
+	}
+
+	fw, err := ffi.New(filePath, &ffi.Config{
 		NodeCacheEntries:     uint(config.CleanCacheSize) / 256, // TODO: estimate 256 bytes per node
 		FreeListCacheEntries: config.FreeListCacheEntries,
 		Revisions:            config.Revisions,
 		ReadCacheStrategy:    config.ReadCacheStrategy,
+		RootStoreDir:         rootStoreDir,
 	})
 	if err != nil {
 		return nil, err
@@ -164,14 +180,14 @@ func (*Database) Scheme() string {
 
 // Initialized checks whether a non-empty genesis block has been written.
 func (db *Database) Initialized(common.Hash) bool {
-	rootBytes, err := db.fwDisk.Root()
+	root, err := db.fwDisk.Root()
 	if err != nil {
 		log.Error("firewood: error getting current root", "error", err)
 		return false
 	}
-	root := common.BytesToHash(rootBytes)
+
 	// If the current root isn't empty, then unless the database is empty, we have a genesis block recorded.
-	return root != types.EmptyRootHash
+	return common.Hash(root) != types.EmptyRootHash
 }
 
 // Update takes a root and a set of keys-values and creates a new proposal.
@@ -320,13 +336,14 @@ func (db *Database) Commit(root common.Hash, report bool) error {
 	defer db.cleanupCommittedProposal(pCtx)
 
 	// Assert that the root of the database matches the committed proposal root.
-	currentRootBytes, err := db.fwDisk.Root()
+	currentRoot, err := db.fwDisk.Root()
 	if err != nil {
 		return fmt.Errorf("firewood: error getting current root after commit: %w", err)
 	}
-	currentRoot := common.BytesToHash(currentRootBytes)
-	if currentRoot != root {
-		return fmt.Errorf("firewood: current root %s does not match expected root %s", currentRoot.Hex(), root.Hex())
+
+	currentRootHash := common.Hash(currentRoot)
+	if currentRootHash != root {
+		return fmt.Errorf("firewood: current root %s does not match expected root %s", currentRootHash.Hex(), root.Hex())
 	}
 
 	if report {
@@ -402,13 +419,14 @@ func createProposal(layer proposable, root common.Hash, keys, values [][]byte) (
 	ffiProposeTimer.Inc(time.Since(start).Milliseconds())
 	ffiOutstandingProposals.Inc(1)
 
-	currentRootBytes, err := p.Root()
+	currentRoot, err := p.Root()
 	if err != nil {
 		return nil, fmt.Errorf("firewood: error getting root of proposal %s: %w", root, err)
 	}
-	currentRoot := common.BytesToHash(currentRootBytes)
-	if root != currentRoot {
-		return nil, fmt.Errorf("firewood: proposed root %s does not match expected root %s", currentRoot.Hex(), root.Hex())
+
+	currentRootHash := common.Hash(currentRoot)
+	if root != currentRootHash {
+		return nil, fmt.Errorf("firewood: proposed root %s does not match expected root %s", currentRootHash.Hex(), root.Hex())
 	}
 
 	return p, nil
@@ -475,16 +493,16 @@ func (db *Database) removeProposalFromMap(pCtx *ProposalContext) {
 // Reader retrieves a node reader belonging to the given state root.
 // An error will be returned if the requested state is not available.
 func (db *Database) Reader(root common.Hash) (database.Reader, error) {
-	if _, err := db.fwDisk.GetFromRoot(root.Bytes(), []byte{}); err != nil {
+	if _, err := db.fwDisk.GetFromRoot(ffi.Hash(root), []byte{}); err != nil {
 		return nil, fmt.Errorf("firewood: unable to retrieve from root %s: %w", root.Hex(), err)
 	}
-	return &reader{db: db, root: root}, nil
+	return &reader{db: db, root: ffi.Hash(root)}, nil
 }
 
 // reader is a state reader of Database which implements the Reader interface.
 type reader struct {
 	db   *Database
-	root common.Hash // The root of the state this reader is reading.
+	root ffi.Hash // The root of the state this reader is reading.
 }
 
 // Node retrieves the trie node with the given node hash. No error will be
@@ -493,7 +511,7 @@ func (reader *reader) Node(_ common.Hash, path []byte, _ common.Hash) ([]byte, e
 	// This function relies on Firewood's internal locking to ensure concurrent reads are safe.
 	// This is safe even if a proposal is being committed concurrently.
 	start := time.Now()
-	result, err := reader.db.fwDisk.GetFromRoot(reader.root.Bytes(), path)
+	result, err := reader.db.fwDisk.GetFromRoot(reader.root, path)
 	if metrics.EnabledExpensive {
 		ffiReadCount.Inc(1)
 		ffiReadTimer.Inc(time.Since(start).Milliseconds())
@@ -540,11 +558,11 @@ func (db *Database) getProposalHash(parentRoot common.Hash, keys, values [][]byt
 	// We succesffuly created a proposal, so we must drop it after use.
 	defer p.Drop()
 
-	rootBytes, err := p.Root()
+	root, err := p.Root()
 	if err != nil {
 		return common.Hash{}, err
 	}
-	return common.BytesToHash(rootBytes), nil
+	return common.Hash(root), nil
 }
 
 func arrangeKeyValuePairs(nodes *trienode.MergedNodeSet) ([][]byte, [][]byte) {